LayoutTests/media/modern-media-controls/volume-slider/volume-slider.html

<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/slider.css" type="text/css" media="screen">
<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/volume-slider.css" type="text/css" media="screen">
<script src="../../../resources/js-test-pre.js"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scheduler.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-node.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/slider.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/volume-slider.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>VolumeSlider</code> class.");

const volumeSlider = new VolumeSlider;

shouldBeTrue('volumeSlider.element.classList.contains("volume")');

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
