LayoutTests/media/modern-media-controls/volume-slider/volume-slider-expected.txt

Testing the VolumeSlider class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS volumeSlider.element.classList.contains("volume") is true
PASS successfullyParsed is true

TEST COMPLETE
